Minor cosmetic changes incl a few more bits of info in java properties info window
git-svn-id: svn://neuroconstruct.org:3999/neuroConstruct/trunk@543 cb8a3d95-5132-0410-b650-8d485996f008

+++ src/ucl/physiol/neuroconstruct/cell/converters/MorphMLConverter.java
@@ -731,7 +731,7 @@ public class MorphMLConverter extends FormatImporter
                             }
 
                             revPotParamElement.addAttribute(new SimpleXMLAttribute(BiophysicsConstants.PARAMETER_VALUE_ATTR,
-                                                                             UnitConverter.getVoltage(revPot,
+                                                                             (float)UnitConverter.getVoltage(revPot,
                                 UnitConverter.NEUROCONSTRUCT_UNITS,
                                 UnitConverter.GENESIS_PHYSIOLOGICAL_UNITS) + ""));
